+++ app/mailers/user_mailer.rb
@@ -18,6 +18,22 @@ class UserMailer < ActionMailer::Base
          subject: "#{user.name}'s Account Activated")
   end
 
+  def user_added_to_project(project_user)
+    setup_email
+    @project_user = project_user
+    mail(to: project_user.user.email,
+         subject: "#{project_user.creator.name} Allows You to View Project #{project_user.project.name}",
+         reply_to: project_user.creator.email)
+  end
+
+  def invite_user_to_project(project_user)
+    setup_email
+    @project_user = project_user
+    mail(to: project_user.invite_email,
+         subject: "#{project_user.creator.name} Invites You to View Project #{project_user.project.name}",
+         reply_to: project_user.creator.email)
+  end
+
   protected
 
   def setup_email